#27: Fixing the duplicating cli
