+++ EpubCore.Cli/CliErrorHandler.cs
@@ -6,9 +6,5 @@ public class CliErrorHandler : ICliErrorHandler
 {
     public void HandleError(IEnumerable<Error> errors)
     {
-        //foreach (var error in errors)
-        //{
-        //    Console.WriteLine(error.);
-        //}
     }
 }
