+++ EpubCore.Cli/EpubCore.Cli.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
 	  <PackAsTool>true</PackAsTool>
@@ -13,7 +13,7 @@
 	  <Copyright>Copyright 2023</Copyright>
 	  <PackageProjectUrl>https://github.com/domingoladron/EpubCore</PackageProjectUrl>
 	  <RepositoryUrl>https://github.com/domingoladron/EpubCore</RepositoryUrl>
-	  <Version>0.8.0.0</Version>
+	  <Version>0.9.0.0</Version>
   </PropertyGroup>
 
   <ItemGroup>
