+++ EpubCore.Cli/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "EpubCore.Cli": {
       "commandName": "Project",
-      "commandLineArgs": "gen-cli-docs --out docs/wiki/epub-cli-docs.md"
+      "commandLineArgs": "--version"
     }
   }
 }
